+++ hydromt/data_catalog/uri_resolvers/resolver_plugin.py
@@ -3,11 +3,11 @@
 from typing import Dict, Type
 
 from .convention_resolver import ConventionResolver
-from .metadata_resolver import MetaDataResolver
 from .raster_tindex_resolver import RasterTindexResolver
+from .uri_resolver import URIResolver
 
 # placeholder for proper plugin behaviour later on.
-RESOLVERS: Dict[str, Type[MetaDataResolver]] = {
+RESOLVERS: Dict[str, Type[URIResolver]] = {
     ConventionResolver.name: ConventionResolver,
     RasterTindexResolver.name: RasterTindexResolver,
 }